Authenticate for Learning Objective CRUD, BaseEntity for Entities,...

See merge request hcm24_cpl_java_01/nhom_2!13

+++ src/main/java/com/example/fams/config/ResponseUtil.java
@@ -21,7 +21,7 @@ public class ResponseUtil {
     }
 
     public static ResponseEntity<?> getCollection(Object result, HttpStatus status, String response
-    ,int page, int limit, int count) {
+    ,int page, int limit, long count) {
         return new ResponseEntity<>(
                 ResponseDTO.builder()
                         .statusCode(status.value())
@@ -33,7 +33,7 @@ public class ResponseUtil {
         );
     }
 
-    public MeatadataDTO getMeatadata(int page, int limit, int count){
+    public MeatadataDTO getMeatadata(int page, int limit, long count){
         MeatadataDTO result = new MeatadataDTO();
         result.setPage(page);
         result.setTotal(((int) Math.ceil((double) count/limit)));
